+++ database/factories/Repositories/UsersRepositoryFactory.php
@@ -16,6 +16,7 @@ class UsersRepositoryFactory extends Factory
     {
         return [
             'email' => 'info@example.com',
+            'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
             'role_id' => 1,
         ];
     }